mark fs as float, suggested in #1

+++ src/pymultitaper/spectral.py
@@ -27,7 +27,7 @@ def _get_1d_window(window_shape,n_winlen):
     win_arr = win_arr[:,None]
     return win_arr,weights
 
-def _spectrogram(data:NDArray,fs:int,time_step:float,win:NDArray,weights:NDArray,freq_range:list,detrend:Literal["constant","linear","off"],nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5) -> Tuple[NDArray,NDArray,NDArray]:
+def _spectrogram(data:NDArray,fs:float,time_step:float,win:NDArray,weights:NDArray,freq_range:list,detrend:Literal["constant","linear","off"],nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5) -> Tuple[NDArray,NDArray,NDArray]:
     """
     Core implementation of spectrogram (PSD) calculation.
 
@@ -35,7 +35,7 @@ def _spectrogram(data:NDArray,fs:int,time_step:float,win:NDArray,weights:NDArray
 
     Args:
         data (NDArray): (n_samples,)
-        fs (int): Sampling frequency
+        fs (float): Sampling frequency
         time_step (float): Time step between frames in seconds
         win (NDArray): (n_winlen,n_wins) Window arrays from different window functions
         weights (NDArray): (n_wins,) Weights for each window array
@@ -104,13 +104,13 @@ def _spectrogram(data:NDArray,fs:int,time_step:float,win:NDArray,weights:NDArray
     return times,freqs,psd_data
 
 
-def multitaper_spectrogram(data:NDArray,fs:int,time_step:float,window_length:float,NW:float=4.0,n_tapers:Optional[int]=None,freq_range:Optional[list]=None,weight_type:Literal["unity","eig"]="unity",detrend:Literal["constant","linear","off"]="constant",nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5)-> Tuple[NDArray,NDArray,NDArray]:
+def multitaper_spectrogram(data:NDArray,fs:float,time_step:float,window_length:float,NW:float=4.0,n_tapers:Optional[int]=None,freq_range:Optional[list]=None,weight_type:Literal["unity","eig"]="unity",detrend:Literal["constant","linear","off"]="constant",nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5)-> Tuple[NDArray,NDArray,NDArray]:
     """
     Compute the multitaper PSD of the input data.
 
     Args:
         data (NDArray): (n_samples,) Input data
-        fs (int): Sampling frequency
+        fs (float): Sampling frequency
         time_step (float): Time step between frames in seconds
         window_length (float): Window length in seconds
         NW (float, optional): NW value, see notes for details. Defaults to 4.0.
@@ -141,13 +141,13 @@ def multitaper_spectrogram(data:NDArray,fs:int,time_step:float,window_length:flo
     tapers,weights = _get_dpss_windows(n_winlen,NW,n_tapers,weight_type)
     return _spectrogram(data=data,fs=fs,time_step=time_step,win=tapers,weights=weights,freq_range=freq_range,detrend=detrend,nfft=nfft,db_scale=db_scale,p_ref=p_ref)
 
-def spectrogram(data:NDArray,fs:int,time_step:float,window_length:float,window_shape:Union[str,tuple]="hamming",freq_range:Optional[list]=None,detrend:Literal["constant","linear","off"]="constant",nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5)-> Tuple[NDArray,NDArray,NDArray]:
+def spectrogram(data:NDArray,fs:float,time_step:float,window_length:float,window_shape:Union[str,tuple]="hamming",freq_range:Optional[list]=None,detrend:Literal["constant","linear","off"]="constant",nfft:Optional[int]=None,db_scale:bool=True,p_ref:float=2e-5)-> Tuple[NDArray,NDArray,NDArray]:
     """
     Compute the ordinary (single-taper) PSD of the input data. This is similar to `scipy.signal.spectrogram`.
 
     Args:
         data (NDArray): (n_samples,) Input data
-        fs (int): Sampling frequency
+        fs (float): Sampling frequency
         time_step (float): Time step between frames in seconds
         window_length (float): Window length in seconds
         window_shape (Union[str,tuple], optional): The shape of the window function. Defaults to "hamming".
